🔥[MODIFY][backend][이홍열] admin 동네정보 게시물 관리 API

+++ backend/backOffice/services/cmpostService.js
@@ -4,7 +4,15 @@ const getCmpost = () => {
   return cmpostDao.getCmpost();
 };
 
-const deleteCmpost = (cmpostId) => {
+const deleteCmpost = async (cmpostId) => {
+  const checkCmpostId = await cmpostDao.checkRegistCmpostId(cmpostId);
+
+  if (!checkCmpostId) {
+    const error = new Error(`NOT_EXIST_COMMUNITY_POST`);
+    error.statusCode = 400;
+    throw error;
+  }
+
   return cmpostDao.deleteCmpost(cmpostId);
 };
